--- src/main/java/raisetech/StudentManagement/Student2.java
@@ -1,32 +1,0 @@
-package raisetech.StudentManagement;
-
-public class Student2 {
-
-  private String name;
-  private int age;
-  private int howLong;
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public int getAge() {
-    return age;
-  }
-
-  public void setAge(int age) {
-    this.age = age;
-  }
-
-  public int getHowLong() {
-    return howLong;
-  }
-
-  public void setHowLong(int howLong) {
-    this.howLong = howLong;
-  }
-}
