+++ src/main/java/raisetech/StudentManagement/data/StudentCourses.java
@@ -1,4 +1,4 @@
-package raisetech.StudentManagement;
+package raisetech.StudentManagement.data;
 
 import java.time.LocalDate;
 import lombok.Getter;
