+++ src/main/java/raisetech/StudentManagement/repository/StudentRepository.java
@@ -1,8 +1,10 @@
-package raisetech.StudentManagement;
+package raisetech.StudentManagement.repository;
 
 import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Select;
+import raisetech.StudentManagement.data.Student;
+import raisetech.StudentManagement.data.StudentCourses;
 
 @Mapper
 public interface StudentRepository {
